+++ modules/terraform-zscc-function-app-azure/variables.tf
@@ -122,3 +122,18 @@ variable "path_to_scripts" {
   description = "Path to script_directory"
   default     = ""
 }
+
+variable "asp_sku_name" {
+  type        = string
+  description = "SKU Name for the App Service Plan. Recommended Y1 (flex consumption) for function app unless not supported by Azure region"
+  default     = "Y1"
+  validation {
+    condition = (
+      var.asp_sku_name == "Y1" ||
+      var.asp_sku_name == "FC1" ||
+      var.asp_sku_name == "EP1" ||
+      var.asp_sku_name == "B1"
+    )
+    error_message = "Input asp_sku_name selected is not a valid/approved SKU Name."
+  }
+}